Split test_jokes
